pyproject.toml
[project]
name = "zk-streamlit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "kazoo>=2.10.0",
    "python-dotenv>=1.0.1",
    "streamlit>=1.39.0",
]
